tasks 4.15-4.23 complete

+++ Osa4/blogilista/controllers/blogs.js
@@ -1,8 +1,13 @@
+const jwt = require('jsonwebtoken')
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const { userExtractor } = require('../utils/middleware')
+
+
 
 blogsRouter.get('/', async (request, response) => {
-    const blogs = await Blog.find({})
+    const blogs = await Blog.find({}).populate('user', { username: 1, name:1, id: 1 })
     response.json(blogs)
 })
 
@@ -15,19 +20,34 @@ blogsRouter.get('/:id', async (request, response) => {
     }
 })
 
-blogsRouter.post('/', async (request, response) => {
-    const body = request.body
+blogsRouter.post('/', userExtractor, async (request, response) => {
+    const body = await request.body
+    const decodedToken = jwt.verify(request.token, 
+    process.env.SECRET)
+    if (!decodedToken) {
+        return response.status(401).json({ error: 'token invalid'})
+    }
+    const user = request.user
     const blog = new Blog({
         title: body.title,
         author: body.author,
         url: body.url,
-        likes: body.likes
+        user: user.id,
+        likes: body.likes,
     })
     const savedBlog = await blog.save()
+    user.blogs = user.blogs.concat(savedBlog._id)
+    await user.save()
     response.status(201).json(savedBlog)
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    const blog = await Blog.findById(request.params.id)
+
+    if (blog.user.toSting() !== decodedToken.id) {
+        return response.status(400).json({ error: 'no auth'})
+    }
     await Blog.findByIdAndDelete(request.params.id)
     response.status(204).end()
 })
